chore(plugins/plugin-tekton): code cleanup for plugin-tekton
Fixes #1489
